Added styling for user interface via xml
Android Studio has 2 ways to add styling
to user interface: one is to drag and drop
GUI components into the user interface to design it.
two is to directly edit the activity_main.xml to
style the user interface. Added text view
for indicating how much money one has, 1 button
to increment the money and 1 button for showing
what is happening on the android app screen.