Revamped if stmt with switch case in makeItRain()
Previously used if statement when moneyCounter reached
10000 to then set the text color. However since my task
was to change the text color for 10000, 30000 and then
50000, a switch case looked easier to read. I also showed
two approaches for setting the text color. First approach
is to call on predefined color from color state list. Second
approach is to get our own hex color resource we defined
in colors.xml.